==> code/02 Data cleaning and formating.Rmd <==
---
title: "Data cleaning and formating"
output: html_notebook
---

# Data sources

The data used in this study are from:

  * **collisions data** from [Road Accident and Safety Statistics Table RAS50016](https://www.gov.uk/government/statistical-data-sets/ras50-contributory-factors)
    and (for 2010 and 2011) from the [National Archives](http://webarchive.nationalarchives.gov.uk/20131210232105/https://www.gov.uk/government/statistical-data-sets/ras50-contributory-factors)
  * **police workforce data** and **police functions data** from [HMICFRS Value for Money data](https://www.justiceinspectorates.gov.uk/hmicfrs/data/value-for-money-data/)
  * **historical police workforce data**
  * **fixed penalty notice data** and **breath test data** from [Police powers and procedures England and Wales statistics](https://www.gov.uk/government/collections/police-powers-and-procedures-england-and-wales)
  * **vehicle kilometres data** from [Road Traffic Statistics Table TRA8904](https://www.gov.uk/government/statistical-data-sets/tra89-traffic-by-local-authority) 

All data are available under [version 3 of the Open Government Licence](https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/). 
This licence explicitly allows relicensing of further-processed data under a 
CC-BY licence.

## Subsidiary data

Other data needed for this project:

  * [Lookup table between police-force areas and local authorities](http://geoportal.statistics.gov.uk/datasets/local-authority-district-to-community-safety-partnerships-to-police-force-areas-december-2016-lookup-in-england-and-wales)


# Pre-processing

Since the government typically does not provide data tables in machine-readable 
formats, the files have been lightly pre-processed before importing into R.
This includes removing footnotes and multiple header rows, as well as separating
multi-worksheet spreadsheets into flat CSV files. All other processing (e.g.
wide-to-long conversion) is done in this file.


# Police workforce and functions data

These data come from HMICFRS, which produces data on both police workforce and 
officer functions. The functions data are consistent over time, whereas the 
corresponding Home Office data used different definitions from 2015 onwards
(see the [User Guide to Police Workforce Statistics](https://www.gov.uk/government/publications/user-guide-to-police-workforce-statistics)). Officers are considered to be allocated to traffic if they appear in the
'Roads Policing' category of the HMICFRS data.

One problem with the HMICFRS data are that they are produced in a slightly
different table format *almost every year*. The data have been pre-processed to
convert the assortment of CSV, Excel and ODS files on their website to a series
of flat CSV files. These files are not [tidy](https://cran.r-project.org/web/packages/tidyr/vignettes/tidy-data.html), 
so some individual wrangling will be required.

Some of the HMICFRS files contain many different variables, and in some cases
the manual pre-processing has to be done for each variable. To minimise this
work (and any consequent mistakes), the only variables that are needed are:

  * police force area population,
  * number of police officers, and
  * workforce by function.

Passive provision should be made to include all functions, just in case they are 
needed for calculating the number of uniformed/front-line officers. Some years 
of data don't include the detailed functions, so only high-level functions (e.g. 
'roads policing') need be included.

HMICFRS uses data from the year prior to the report, so the data listed on their 
website as from 2010 is actually data from 2009.

```{r}
hmic <- list()

# 2010 data
hmic[["2010"]] <- read_csv("../original_data/hmic_data_2011.csv") %>% 
  select(force = "Police force", 
         officers = "Home Office published data 2009-10_FTE equivalent staff_Police Officers",
         roads_officers = "Workforce by function FTE_Police Officers (as at 31/03/2010)_Roads policing") %>% 
  mutate(force = harmonise_force(force), year = 2010)

# 2011 data
hmic[["2011"]] <- read_csv("../original_data/hmic_data_2012.csv") %>% 
  rename(officer_function = Function) %>% 
  rename_all(tolower) %>% 
  mutate(officer_function = tolower(officer_function)) %>% 
  filter(officer_function %in% c("total road policing", "level 2 summary")) %>% 
  gather("force", "value", -officer_function) %>% 
  spread(officer_function, value) %>% 
  rename(officers = `level 2 summary`, 
         roads_officers = `total road policing`) %>% 
  mutate(force = harmonise_force(force), year = 2011)

# 2012—2016 data
# note that the 2016 data have an asterisk after some category names, which have
# been removed during pre-processing so that the file can be processed along
# with the others
hmic_list <- lapply(2012:2016, function (x) {
  read_csv(paste0("../original_data/hmic_data_", (x + 1), ".csv")) %>% 
    rename_all(tolower) %>% 
    rename(poa_category = `poa category`) %>% 
    mutate(poa_category = tolower(poa_category)) %>% 
    filter(poa_category %in% c("total road policing", "level 2 summary")) %>%
    gather("key", "value", -poa_category) %>%
    spread(poa_category, value) %>%
    rename(force = key, officers = `level 2 summary`,
           roads_officers = `total road policing`) %>%
    mutate(force = harmonise_force(force), year = x)
})

# merge the individual years with the combined years
bind_rows(c(hmic, hmic_list)) %>% 
  write_csv("../analysis_data/data_officer_counts.csv", na = "")
rm(hmic, hmic_list)

```


# Historical police workforce data

These data are needed for a chart of the number of police per head since the 
Second World War.


# Historical population data


# FPN data

Data on FPNs issued from 2011 onwards are available by year in the 2017 Home
Office police powers statistics. 2010 data is available as a separate file,
which also includes data back to 2007 (although camera FPN counts are only
available from 2009).

The notes in the 2017 FPN file state that all Suffolk FPNs are issued jointly
with Norfolk, so these forces should be combined for analysis. A similar issue
affects South Wales, Gwent and Dyfed-Powys, which should also be combined in the
final analysis.

First we'll deal with the 2007—2010 data file.

```{r}
fpn_data <- list()

# load 2007—2010 data
fpn_data_raw <- read_csv("../original_data/fpn_data_2007-2010.csv") %>% 
  select(year = Year, force = `Police force`, offence = Offence, 
         fpns = `Total FPNs`) %>% 
  # totals in the dataset are prefixed with an *, so we can remove these rows
  filter(!str_detect(force, "^\\*")) %>% 
  mutate(force = harmonise_force(force))

# count all FPNs
fpn_all <- fpn_data_raw %>% 
  filter(offence == "*All offences") %>% 
  group_by(year, force) %>% 
  summarise(fpn_all = sum(fpns))

# count mobile phone FPNs
fpn_phone <- fpn_data_raw %>% 
  filter(offence == "Use of hand-held mobile phone while driving") %>% 
  group_by(year, force) %>% 
  summarise(fpn_phone = sum(fpns))

# count speeding FPNs
fpn_speed <- fpn_data_raw %>% 
  filter(offence == "Speed limit offences") %>% 
  group_by(year, force) %>% 
  summarise(fpn_speed = sum(fpns))

# get camera offence counts
fpn_data_camera <- read_csv("../original_data/fpn_data_2007–2010_camera.csv") %>% 
  select(year = Year, force = `Police force`, offence = `Offence`, 
         fpns = `Camera detected FPNS`) %>% 
  # totals in the dataset are prefixed with an *, so we can remove these rows
  filter(!str_detect(force, "^\\*") & year >= 2009 & 
         offence != "Total camera detections") %>% 
  mutate(force = harmonise_force(force)) %>%
  spread(offence, fpns) %>%
  rename(fpn_camera_speed = `Speed limit offences`,
         fpn_camera_other = `Neglect of traffic directions`)

# merge these datasets
fpn_data[["2010"]] <- fpn_all %>% 
  left_join(fpn_phone, by = c("year" = "year", "force" = "force")) %>% 
  left_join(fpn_speed, by = c("year" = "year", "force" = "force")) %>% 
  left_join(fpn_data_camera, by = c("year" = "year", "force" = "force"))

rm(fpn_data_raw, fpn_all, fpn_phone, fpn_speed, fpn_data_camera)
```

Now we can deal with the 2011—2016 data and merge the two files.

```{r}
# load data
fpn_data_raw <- read_csv("../original_data/fpn_data_2011–2017.csv") %>% 
  select(year = `Calendar Year`, force = `Police Force Area`, 
         offence = `Offence Description`, outcome = Outcome, 
         fpns = `Number of FPNs`) %>% 
  filter(outcome != "Cancelled") %>% 
  mutate(force = harmonise_force(force))

# count all FPNs
fpn_all <- fpn_data_raw %>% 
  group_by(year, force) %>% 
  summarise(fpn_all = sum(fpns))

# count mobile phone FPNs
fpn_phone <- fpn_data_raw %>% 
  filter(offence == "Use of hand-held mobile phone while driving") %>% 
  group_by(year, force) %>% 
  summarise(fpn_phone = sum(fpns))

# count speeding FPNs
fpn_speed <- fpn_data_raw %>% 
  filter(offence == "Speed limit offences") %>% 
  group_by(year, force) %>% 
  summarise(fpn_speed = sum(fpns))

# get camera offence counts
fpn_data_camera <- read_csv("../original_data/fpn_data_2011–2017_camera.csv") %>% 
  select(year = `Calendar Year`, force = `Police Force Area`, 
         offence = `Offence Description`, type = `Detection type`, 
         fpns = `Number of FPNs`) %>% 
  mutate(force = harmonise_force(force)) %>% 
  filter(type == "Camera-detected") %>% 
  group_by(year, force, offence) %>% 
  summarise(fpn_camera = sum(fpns)) %>% 
  ungroup() %>% 
  spread(offence, fpn_camera) %>% 
  rename(fpn_camera_speed = `Speed limit offences`,
         fpn_camera_other = `Neglect of traffic directions`)

# merge these datasets
fpn_data[["2016"]] <- fpn_all %>% 
  left_join(fpn_phone, by = c("year" = "year", "force" = "force")) %>% 
  left_join(fpn_speed, by = c("year" = "year", "force" = "force")) %>% 
  left_join(fpn_data_camera, by = c("year" = "year", "force" = "force"))

rm(fpn_data_raw, fpn_all, fpn_phone, fpn_speed, fpn_data_camera)

# merge the two data files together
bind_rows(fpn_data) %>%
  arrange(year, force) %>%
  filter(
    year >= 2009 & 
    !force %in% c("Dyfed-Powys", "Gwent", "North Wales", "South Wales")
  ) %>% 
  mutate(
    fpn_camera = fpn_camera_speed + fpn_camera_other,
    fpn_all = fpn_all - fpn_camera,
    fpn_speed = fpn_speed - fpn_camera_speed
  ) %>%
  select(-fpn_camera_other) %>%
  write_csv("../analysis_data/data_fpn.csv", na = "")

rm(fpn_data)
```


# Breath test data

Monthly breath test data is published by the Home Office in a single file from 
2002 onwards. The data are in tidy format already, so they can just be
aggregated to annual totals and outputted again.

```{r}
read_csv("../original_data/breath_test_data.csv") %>% 
  select(year = `Calendar Year`, force = `Police Force`, 
         breath_tests = `Number of breath tests`) %>% 
  mutate(force = harmonise_force(force)) %>% 
  filter(year >= 2009) %>% 
  group_by(year, force) %>% 
  summarise(breath_tests = sum(breath_tests)) %>% 
  arrange(year, force) %>% 
  write_csv("../analysis_data/data_breath_tests.csv", na = "")
```


# Collision data

Collision data from 2010 onwards are available from the gov.uk website as annual
counts at top-tier local-authority level. We will need to aggregate these to PFA 
level. *Collision counts are available only for England and Wales.*

These estimates can be converted to police-force areas (PFAs), except that the 
ONS lookup table for PFAs only links to local-authority districts. This is fine
for unitary authorities, but not for two-tier areas. To get around this we can
create a lookup using the ONS lookup file for counties and districts.

## Counts of collisions

The collision counts have multiple header rows that have been merged during
pre-processing. "London Airport (Heathrow)" is listed as a separate local
authority without an ONS code, so during pre-processing it has been given the
local authority code for Hounslow (E09000018).

```{r}
# load LAD->PFA lookup table
pfa_lad_lookup <- read_csv("../analysis_data/lookup_pfa_to_lad.csv") %>% 
  select(LAD16CD, force = PFA16NM) %>% 
  mutate(force = str_replace_all(force, " & ", " and "))

# create CTY->PFA lookup table
pfa_cty_lookup <- read_csv("../analysis_data/lookup_cty_to_lad.csv") %>% 
  left_join(pfa_lad_lookup, by = c("LAD16CD" = "LAD16CD")) %>% 
  group_by(CTY16CD) %>%
  summarise(force = first(force))

# load and process collision counts
collision_counts <- lapply(2010:2016, function (x) {
  read_csv(paste0("../original_data/collision_counts_", x, ".csv")) %>% 
    select(la_code = `ONS Code`, acc_killed = `Fatal`, acc_serious = `Serious`,
           acc_all = `All`) %>% 
    mutate(
      acc_ksi = acc_killed + acc_serious,
      la_code = if_else(la_code == "E06000048", "E06000057", la_code),
      la_code = if_else(la_code == "E08000020", "E08000037", la_code),
      year = x
    ) %>% 
    select(-acc_killed, -acc_serious) %>% 
    left_join(pfa_lad_lookup, by = c("la_code" = "LAD16CD")) %>% 
    left_join(pfa_cty_lookup, by = c("la_code" = "CTY16CD")) %>% 
    mutate(force = if_else(is.na(force.x), force.y, force.x)) %>% 
    filter(!is.na(force)) %>% 
    group_by(year, force) %>% 
    summarise(acc_all = sum(acc_all), acc_ksi = sum(acc_ksi))
}) %>% bind_rows() %>% 
  mutate(force = harmonise_force(force))
```


## Collision data by contributory factors

For some years, some of the variable names have numbers at the end to indicate
footnotes. These have been removed during pre-processing.

```{r}
# load and process the collisions data files
collision_factor_data <- lapply(2010:2016, function (x) {
  read_csv(paste0("../original_data/collisions_data_", x, ".csv")) %>% 
    select(la_code = `ONS Code`,
           acc_speeding = `Exceeding speed limit`, 
           acc_drink = `Driver/Rider impaired by alcohol`, 
           acc_phone = `Driver using mobile phone`) %>% 
    mutate(
      la_code = if_else(la_code == "E06000048", "E06000057", la_code),
      la_code = if_else(la_code == "E08000020", "E08000037", la_code)
    ) %>%
    mutate(year = x) %>% 
    left_join(pfa_lad_lookup, by = c("la_code" = "LAD16CD")) %>% 
    left_join(pfa_cty_lookup, by = c("la_code" = "CTY16CD")) %>% 
    mutate(force = if_else(is.na(force.x), force.y, force.x)) %>% 
    filter(!is.na(force)) %>% 
    group_by(year, force) %>% 
    summarise(acc_speeding = sum(acc_speeding), acc_drink = sum(acc_drink),
              acc_phone = sum(acc_phone))
}) %>% bind_rows() %>% 
  mutate(force = harmonise_force(force))
```

## Combining collision datasets

The Collision counts data only covers police forces in England, while the
contributory-factor data covers England and Wales. By using a left join, data
from Wales will be discarded.

```{r}
collision_counts %>% 
  left_join(collision_factor_data, 
            by = c("year" = "year", "force" = "force")) %>% 
  write_csv("../analysis_data/data_collisions.csv", na = "")

rm(collision_counts, collision_factor_data, pfa_cty_lookup, pfa_lad_lookup)
```



# Vehicle kilometres data

Estimates of millions of vehicle-kilometres travelled are published for top-tier
local authorities for each year.

As with the collision data, we will need to create a CTY->LAD lookup to go with
the existing ONS LAD->PFA lookup table.

```{r}
# load traffic data
traffic_data_la <- read_csv("../original_data/traffic_data.csv") %>% 
  gather("year", "million_km", -`LA Code`, -`Local Authority`) %>% 
  mutate(year = as.integer(year)) %>% 
  # the next line corrects two problems in the traffic data file
  mutate(
    `LA Code` = if_else(`LA Code` == "E06000048", "E06000057", `LA Code`),
    `LA Code` = if_else(`LA Code` == "E08000020", "E08000037", `LA Code`)
  ) %>%
  filter(year >= 2009)

# load LAD->PFA lookup table
pfa_lad_lookup <- read_csv("../analysis_data/lookup_pfa_to_lad.csv") %>% 
  select(LAD16CD, force = PFA16NM) %>% 
  mutate(force = str_replace_all(force, " & ", " and "))

# create CTY->PFA lookup table
pfa_cty_lookup <- read_csv("../analysis_data/lookup_cty_to_lad.csv") %>% 
  left_join(pfa_lad_lookup, by = c("LAD16CD" = "LAD16CD")) %>% 
  group_by(CTY16CD) %>%
  summarise(force = first(force))

# join PFA name to traffic data
traffic_data_la <- traffic_data_la %>% 
  left_join(pfa_lad_lookup, by = c("LA Code" = "LAD16CD")) %>% 
  left_join(pfa_cty_lookup, by = c("LA Code" = "CTY16CD")) %>% 
  mutate(force = if_else(is.na(force.x), force.y, force.x)) %>% 
  select(-force.x, -force.y) %>% 
  filter(!is.na(force))

# sum traffic counts for PFAs
traffic_data_pfa <- traffic_data_la %>% 
  group_by(year, force) %>% 
  summarise(million_km = sum(million_km))

# write CSV file
traffic_data_pfa %>% 
  mutate(force = case_when(
    force == "London, City of" ~ "City of London",
    force == "Metropolitan Police" ~ "Metropolitan",
    TRUE ~ force
  )) %>% 
  write_csv("../analysis_data/data_traffic.csv", na = "")

rm(pfa_cty_lookup, pfa_lad_lookup, traffic_data_la, traffic_data_pfa)
```


# Combining the data

Each of the sections above outputs a tidy CSV file. We can now combine these to
produce the final data file that will be used for analysis. At the same time we
can remove the Welsh forces (for which we don't have accident counts) and merge
Norfolk and Suffolk (since they have a single Central Ticket Office).

```{r}
dir("../analysis_data/", pattern = "data_*", full.names = TRUE) %>% 
  map(read_csv) %>% 
  plyr::join_all(by = c("year", "force"), type = "left") %>% 
  filter(
    year >= 2010 &
    !force %in% c("Dyfed-Powys", "Gwent", "North Wales", "South Wales")
  ) %>% 
  mutate_if(is.numeric, as.integer) %>% 
  mutate(force = if_else(force %in% c("Norfolk", "Suffolk"), 
                         "Norfolk/Suffolk", force)) %>% 
  group_by(year, force) %>% 
  summarise_all(sum) %>% 
  write_csv("../analysis_data/traffic_police_and_collisions.csv", na = "")
```
