==> code/03 Data analysis.Rmd <==
---
title: "Data analysis"
output:
  html_notebook: default
---

# Load data and create new variables for analysis

All of the data imported into this file are counts (of officers, FPNs etc.). For
the analysis, some variables need to be converted to different formats:

  * traffic officers should be expressed as a proportion of all officers, and
  * FPNs and breath tests should be expressed per officer.
  
This will allow these variables to be included in models along with the number
of officers in each force.

Since the Met is so much bigger than other forces, it's necessary to convert the
counts of accidents into a rate per 10 million vehicle kilometres travelled. 
This avoids skewing the residuals of the regression models by having such a 
large outlier in the data.

Finally, we need to convert `force` from a character to a factor.

```{r}
d <- read_csv("../analysis_data/traffic_police_and_collisions.csv") %>% 
  mutate(
    roads_officers_prop = roads_officers / officers,
    fpn_all_rate = fpn_all / officers,
    fpn_phone_rate = fpn_phone / officers,
    fpn_speed_rate = fpn_speed / officers,
    breath_test_rate = breath_tests  /officers,
    fpn_camera_rate = fpn_camera / (million_km / 10),
    acc_all_rate = acc_all / (million_km / 10),
    acc_ksi_rate = acc_ksi / (million_km / 10),
    acc_drink_rate = acc_drink / (million_km / 10),
    acc_phone_rate = acc_phone / (million_km / 10),
    acc_speed_rate = acc_speeding / (million_km / 10),
    police_force = as.factor(force)
  )
```

## Centering variables

To ease interpretation of the intercept terms in the models below, we should
mean-center some of the variables. Otherwise the intercept term will be the 
rate of accidents expected if there were no police etc.

```{r}
d <- d %>% mutate(
  officers_mctr = scale(officers, center = TRUE, scale = FALSE),
  roads_officers_prop_mctr = scale(roads_officers_prop, center = TRUE, 
                                   scale = FALSE),
  fpn_all_rate_mctr = scale(fpn_all_rate, center = TRUE, scale = FALSE),
  fpn_phone_rate_mctr = scale(fpn_phone_rate, center = TRUE, scale = FALSE),
  fpn_speed_rate_mctr = scale(fpn_speed_rate, center = TRUE, scale = FALSE),
  fpn_camera_rate_mctr = scale(fpn_camera_rate, center = TRUE, scale = FALSE),
  breath_test_rate_mctr = scale(breath_test_rate, center = TRUE, scale = FALSE)
)
```

## Creating lag variables

We know that there is a correlation structure in our data since accident rates 
are likely to be autocorrelated for each force. We can specify this in the model 
structure, but it is probably easier to simply include an appropriate lag term
as a predictor in the model.

```{r}
d <- d %>% 
  arrange(year) %>% 
  group_by(police_force) %>% 
  mutate(
    acc_all_rate_lag = lag(acc_all_rate),
    acc_ksi_rate_lag = lag(acc_ksi_rate)
  ) %>% 
  ungroup()
```



# Checking model assumptions

OLS regression requires a normally distributed response variable, so first we
should check this. We're primarily interested in KSIs, so this is the data we'll
check.

```{r}
hist(d$acc_ksi_rate, breaks = 100)
```

The outlier variables are all for City of London, so this force will be removed
from the data before analysis.

```{r}
d <- d %>% filter(police_force != "City of London")
```


# Model preparation

This list will store the results of all the models run below.

```{r}
m <- list()
```

We need an empty model to compare our models to. Note that `all_` refers to
models predicting *all* collisions, regardless of severity or contributory
factors, while `ksi_` refers to collisions in which somewas was *Killed or 
Seriously Injured*.

This 'empty' model actually includes the lag term, since this is required for
the structure of the data to be modelled correctly.

```{r}
m[["ksi_empty"]] <- lm(
  acc_ksi_rate ~ acc_ksi_rate_lag, 
  data = d
)
summary(m$ksi_empty)
```

We can also use this model to check if the lag term is doing its job by dealing
with the autocorrelation in the data.

```{r}
acf(residuals(m$ksi_empty, type = "pearson"))
```

We can also use a Durbin-Watson test to check the same thing.

```{r}
lmtest::dwtest(m$ksi_empty, alternative = "greater")
```

The residuals are not significantly auto-correlated, so the lag term (together
with the grouping variable) seems to be doing its job.


# Collisions vs officers

This first model predicts collisions using number of officers.

```{r}
m[["ksi_ofc"]] <- update(m$ksi_empty, . ~ . + police_force + officers_mctr)
summary(m$ksi_ofc)
```

From this we can see that the co-efficient for `officers_mctr` is very small and
not significant. The `police_force` variable is significant, which is to be
expected.

```{r}
anova(m$ksi_empty, m$ksi_ofc)
```

The new model is also better than the empty model.


# Collisions vs officers and traffic officers

We can add the proportion of officers who are traffic officers to the model to 
see if officers with training and experience in traffic matters has any effect.

```{r}
m[["ksi_ofc_tpc"]] <- update(m$ksi_ofc, . ~ . + roads_officers_prop_mctr)
summary(m$ksi_ofc_tpc)
```

The co-efficient for `roads_officers_prop_mctr` is larger than for 
`officers_mctr` but non-significant.

```{r}
anova(m$ksi_empty, m$ksi_ofc, m$ksi_ofc_tpc)
```

As expected (since the co-efficient for the only new variable is not 
significant) the new model is not significantly better than the old one.


# Collisions vs officers, traffic officers and activity

We're interested in two types of traffic policing activity: FPNs issued (by 
officers, not by fixed automatic cameras) and breath tests.

```{r}
m[["ksi_ofc_tpc_act"]] <- update(m$ksi_ofc_tpc, 
  . ~ . + fpn_all_rate_mctr + breath_test_rate_mctr)
summary(m$ksi_ofc_tpc_act)
```

```{r}
anova(m$ksi_empty, m$ksi_ofc, m$ksi_ofc_tpc, m$ksi_ofc_tpc_act)
```


# Collisions vs officers, activity and camera activity

```{r}
m[["ksi_ofc_tpc_act_cam"]] <- update(m$ksi_ofc_tpc_act, 
  . ~ . + fpn_camera_rate_mctr)
summary(m$ksi_ofc_tpc_act_cam)
```

```{r}
anova(m$ksi_empty, m$ksi_ofc_tpc_act, m$ksi_ofc_tpc_act_cam)
```

# Next things to consider

  * The need for an interaction term between `police_force` and `officers_mctr`
    because the effect of more officers might be different in different areas
    — the current models only capture variation in officers for the reference
    force and assume it's the same for all other forces.
    Since `police_force` has `r length(levels(d$police_force))` levels, an 
    interaction term would be unwealdy, so this would necessity categorising
    forces, probably based on how urban they are. Three categories would 
    probably be OK: very urban (Met, GMP, WMP and maybe Merseyside), urban and 
    rural.
  * Simulating the power of the model to detect associations between accidents
    and each of the predictors.
